+++ www/go/core/model/Settings.php
@@ -252,6 +252,14 @@ class Settings extends core\Settings {
 	 * @var int
 	 */
 	public $passwordMinLength = 6;
+
+
+	/**
+	 * Number of seconds to auto logout the user if inactive.
+	 * Also disables the "remember login" feature as it would make no sense.
+	 * @var int
+	 */
+	public $logoutWhenInactive = 0;
 	
 	
 	/**