+++ www/views/Extjs3/javascript/form/HtmlEditor.js
@@ -172,6 +172,8 @@ Ext.extend(GO.form.HtmlEditor, Ext.form.HtmlEditor, {
 
 		});
 
+		go.ActivityWatcher.registerDocument(doc);
+
 	},
 
 	debounceTimeout : null,