+++ www/go/core/Installer.php
@@ -138,6 +138,9 @@ class Installer {
 
 		$database->setUtf8();
 
+		//set DYNAMIC row format by default
+		go()->getDbConnection()->exec("SET GLOBAL innodb_default_row_format=DYNAMIC;");
+
 		Utils::runSQLFile(Environment::get()->getInstallFolder()->getFile("go/core/install/install.sql"));
 		App::get()->getDbConnection()->exec("SET FOREIGN_KEY_CHECKS=0;");
 		
@@ -391,6 +394,8 @@ class Installer {
 		
 		go()->getDbConnection()->delete("core_entity", ['name' => 'GO\\Projects\\Model\\Project'])->execute();
 
+		//set DYNAMIC row format by default
+		go()->getDbConnection()->exec("SET GLOBAL innodb_default_row_format=DYNAMIC;");
 		
 		while (!$this->upgradeModules()) {
 			echo "\n\nA module was refactored. Rerunning...\n\n";