+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Core: Upgrade form 6.3 was broken due to toggleGarbageCollection()
 - Core: New configuration option to logout users when inactive for more than x number of seconds.
 - Core: Raised size of search cache from 190 chars to 750 chars
+- Core: fixed <br /> tags in error messages 
 
 23-10-2020 6.4.184
 - Core: Template parse gave error on arrays